fix: updates plots
+ adds a grid plot showing next three days values
